(feat): add ipc between electron and react
this commit adds a typed interface to allow react and electron to safely communicate
to save videos and grab video urls from a users video folder. Some light refactoring
had to be done concerning how blobs were stored relative to the main component
opting for a ref instead of handling in a function.